feat: implement dynamic formant pitch link in audio playback
* Safely checked out the clean `c00f9ef` branch baseline to resolve the `useAudioEngine.ts` syntax corruption and restore the `audioEngine` refactored architecture.
* Updated `SamplerNoteParams`, `Note`, and `SamplerBankParams` in `src/types.ts` to include `formantPitchLink`, `coarseTune`, and `fineTune` to guarantee TypeScript safety.
* Implemented the DSP calculation for "Dynamic Formant Pitch Link" directly inside the `playSamplerVoice` orchestration wrapper in `samplerPlayback.ts`. This calculates the pitch delta from the sample's root note and shifts the formants proportionally via the `formantPitchLink` ratio, passing an absolute value to the stateless audio engine.
* Re-formatted `useKnobInteraction.ts` layout checks.
* Checked off the completed task in `agent_plan.md` and added "Wavetable TTS Oscillation" to the Innovation Lab.
